+++ NEWS.md
@@ -20,6 +20,8 @@
 
 - Document the HPC limitations of `target(format = "keras")` (#989).
 - Remove the now-superfluous vignette.
+- Wrap up console and text file logging functionality into a reference class (#964).
+- Deprecate the `verbose` argument in various caching functions. The location of the cache is now only printed in `make()`. This made the previous feature easier to implement.
 
 
 # Version 7.6.1